// src/core/referal-member/enum/referal-member-enum.ts
export enum REFERAL_MEMBER_ERROR {
  USER_ALREADY_REFERAL_MEMBER = 'REFERAL_MEMBER_ERROR_USER_ALREADY_REFERAL_MEMBER',
  YOU_ALREADT_REFERAL_MEMBER = 'REFERAL_MEMBER_ERROR_YOU_ALREADT_REFERAL_MEMBER',
  USER_IS_MEMBER_OF_YOUR_COMPANY = 'REFERAL_MEMBER_ERROR_USER_IS_MEMBER_OF_YOUR_COMPANY',
  REFERAL_IS_MEMBER_OF_YOUR_COMPANY = 'REFERAL_MEMBER_ERROR_REFERAL_IS_MEMBER_OF_YOUR_COMPANY',
  CANNOT_ADD_YOURSELF_AS_REFERAL_MEMEBER = 'REFERAL_MEMBER_ERROR_CANNOT_ADD_YOURSELF_AS_REFERAL_MEMEBER',
  CANNOT_ADD_ROLE_ADMIN = 'REFERAL_MEMBER_ERROR_CANNOT_ADD_ROLE_ADMIN',
  CANNOT_JOIN_AS_ADMIN = 'REFERAL_MEMBER_ERROR_CANNOT_JOIN_AS_ADMIN',
}