// src/core/referal-payment/enum/referal-payment-type.enum.ts
export enum REFERAL_PAYMENT_TYPE {
  NEW_REFERAL = 0, // бонус при регистрации реферала
  ADDED_KONTRAGENT_INFO = 1, // бонус при указании доп инфы
  USER_TO_REVISION_KONTRAGENT_PAY = 2, // использован для оплаты проверки контрагента
  USER_TO_REVISION_SELF_PAY = 3, // использован для оплаты проверки своей компании
}